requirements.txt
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
cffi==1.17.1
click==8.1.8
cryptography==44.0.2
Django==5.2
djangorestframework==3.16.0
fastapi==0.115.12
greenlet==3.2.0
h11==0.14.0
idna==3.10
Jinja2==3.1.6
MarkupSafe==3.0.2
pycparser==2.22
pydantic==2.11.3
pydantic_core==2.33.1
sniffio==1.3.1
SQLAlchemy==2.0.40
sqlparse==0.5.3
starlette==0.46.2
typing-inspection==0.4.0
typing_extensions==4.13.2
uvicorn==0.34.2
